# Swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API Documentation",
      "description": "API documentation for the provided services across five main modules: Projects, Notifications, Password Reset Email, Tickets, and Users."
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
      "http",
      "https"
    ],
    "paths": {
...
      "/notifications": {
        "get": {
          "tags": [
            "Notifications"
          ],
          "summary": "Get notifications",
          "description": "Retrieve all notifications, optionally filtered by user ID",
          "parameters": [
            {
              "name": "id_usuario",
              "in": "query",
              "type": "integer",
              "description": "Filter notifications by user ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Notifications retrieved successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Notification"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Notifications"
          ],
          "summary": "Create a notification",
          "description": "Creates a new notification for a user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/CreateNotification"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Notification created successfully",
              "schema": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Notification": {
        "type": "object",
        "properties": {
          "id_notificacion": {
            "type": "integer"
          },
          "id_usuario": {
            "type": "integer"
          },
          "mensaje": {
            "type": "string"
          },
          "id_tipo": {
            "type": "integer"
          },
          "fecha_envio": {
            "type": "string",
            "format": "date-time"
          },
          "leido": {
            "type": "boolean"
          }
        }
      },
      "CreateNotification": {
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "integer"
          },
          "mensaje": {
            "type": "string"
          },
          "id_tipo": {
            "type": "integer"
          }
        },
        "required": [
          "id_usuario",
          "mensaje",
          "id_tipo"
        ]
      },
...
}}}